DONE: clean architecture, reorganize backend code
