+++ web/templates/index.html
@@ -11,7 +11,7 @@
 		<h1>Bem-vindo ao Banco Galáxia!</h1>
 		<div class="row">
 			<div class="col">
-				<h3>Usuários</h3>
+				<h3>Usuários</h3>	
 				<div id="usersTable"></div>
 			</div>
 			<div class="col border-l" id="accountInfo" hidden="true">
